+++ test/unit/type1_parser_spec.js
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-import { SEAC_ANALYSIS_ENABLED } from "../../src/core/fonts.js";
+import { SEAC_ANALYSIS_ENABLED } from "../../src/core/fonts_utils.js";
 import { StringStream } from "../../src/core/stream.js";
 import { Type1Parser } from "../../src/core/type1_parser.js";
